# server/src/swagger_server/models/auth_response.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server.models.response import Response  # noqa: F401,E501
from swagger_server import util


class AuthResponse(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, success: bool=None, user_id: int=None, is_admin: bool=None, user_full_name: str=None, auth_token: str=None, token_expiration_datetime: datetime=None):  # noqa: E501
        """AuthResponse - a model defined in Swagger

        :param success: The success of this AuthResponse.  # noqa: E501
        :type success: bool
        :param user_id: The user_id of this AuthResponse.  # noqa: E501
        :type user_id: int
        :param is_admin: The is_admin of this AuthResponse.  # noqa: E501
        :type is_admin: bool
        :param user_full_name: The user_full_name of this AuthResponse.  # noqa: E501
        :type user_full_name: str
        :param auth_token: The auth_token of this AuthResponse.  # noqa: E501
        :type auth_token: str
        :param token_expiration_datetime: The token_expiration_datetime of this AuthResponse.  # noqa: E501
        :type token_expiration_datetime: datetime
        """
        self.swagger_types = {
            'success': bool,
            'user_id': int,
            'is_admin': bool,
            'user_full_name': str,
            'auth_token': str,
            'token_expiration_datetime': datetime
        }

        self.attribute_map = {
            'success': 'success',
            'user_id': 'userId',
            'is_admin': 'isAdmin',
            'user_full_name': 'userFullName',
            'auth_token': 'authToken',
            'token_expiration_datetime': 'tokenExpirationDatetime'
        }
        self._success = success
        self._user_id = user_id
        self._is_admin = is_admin
        self._user_full_name = user_full_name
        self._auth_token = auth_token
        self._token_expiration_datetime = token_expiration_datetime

    @classmethod
    def from_dict(cls, dikt) -> 'AuthResponse':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The AuthResponse of this AuthResponse.  # noqa: E501
        :rtype: AuthResponse
        """
        return util.deserialize_model(dikt, cls)

    @property
    def success(self) -> bool:
        """Gets the success of this AuthResponse.


        :return: The success of this AuthResponse.
        :rtype: bool
        """
        return self._success

    @success.setter
    def success(self, success: bool):
        """Sets the success of this AuthResponse.


        :param success: The success of this AuthResponse.
        :type success: bool
        """

        self._success = success

    @property
    def user_id(self) -> int:
        """Gets the user_id of this AuthResponse.


        :return: The user_id of this AuthResponse.
        :rtype: int
        """
        return self._user_id

    @user_id.setter
    def user_id(self, user_id: int):
        """Sets the user_id of this AuthResponse.


        :param user_id: The user_id of this AuthResponse.
        :type user_id: int
        """

        self._user_id = user_id

    @property
    def is_admin(self) -> bool:
        """Gets the is_admin of this AuthResponse.


        :return: The is_admin of this AuthResponse.
        :rtype: bool
        """
        return self._is_admin

    @is_admin.setter
    def is_admin(self, is_admin: bool):
        """Sets the is_admin of this AuthResponse.


        :param is_admin: The is_admin of this AuthResponse.
        :type is_admin: bool
        """

        self._is_admin = is_admin

    @property
    def user_full_name(self) -> str:
        """Gets the user_full_name of this AuthResponse.


        :return: The user_full_name of this AuthResponse.
        :rtype: str
        """
        return self._user_full_name

    @user_full_name.setter
    def user_full_name(self, user_full_name: str):
        """Sets the user_full_name of this AuthResponse.


        :param user_full_name: The user_full_name of this AuthResponse.
        :type user_full_name: str
        """

        self._user_full_name = user_full_name

    @property
    def auth_token(self) -> str:
        """Gets the auth_token of this AuthResponse.


        :return: The auth_token of this AuthResponse.
        :rtype: str
        """
        return self._auth_token

    @auth_token.setter
    def auth_token(self, auth_token: str):
        """Sets the auth_token of this AuthResponse.


        :param auth_token: The auth_token of this AuthResponse.
        :type auth_token: str
        """

        self._auth_token = auth_token

    @property
    def token_expiration_datetime(self) -> datetime:
        """Gets the token_expiration_datetime of this AuthResponse.


        :return: The token_expiration_datetime of this AuthResponse.
        :rtype: datetime
        """
        return self._token_expiration_datetime

    @token_expiration_datetime.setter
    def token_expiration_datetime(self, token_expiration_datetime: datetime):
        """Sets the token_expiration_datetime of this AuthResponse.


        :param token_expiration_datetime: The token_expiration_datetime of this AuthResponse.
        :type token_expiration_datetime: datetime
        """

        self._token_expiration_datetime = token_expiration_datetime