# server/src/swagger_server/models/user_with_password.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server.models.user import User  # noqa: F401,E501
from swagger_server import util


class UserWithPassword(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, user_id: int=None, username: str=None, is_admin: bool=None, full_name: str=None, password: str=None):  # noqa: E501
        """UserWithPassword - a model defined in Swagger

        :param user_id: The user_id of this UserWithPassword.  # noqa: E501
        :type user_id: int
        :param username: The username of this UserWithPassword.  # noqa: E501
        :type username: str
        :param is_admin: The is_admin of this UserWithPassword.  # noqa: E501
        :type is_admin: bool
        :param full_name: The full_name of this UserWithPassword.  # noqa: E501
        :type full_name: str
        :param password: The password of this UserWithPassword.  # noqa: E501
        :type password: str
        """
        self.swagger_types = {
            'user_id': int,
            'username': str,
            'is_admin': bool,
            'full_name': str,
            'password': str
        }

        self.attribute_map = {
            'user_id': 'userId',
            'username': 'username',
            'is_admin': 'isAdmin',
            'full_name': 'fullName',
            'password': 'password'
        }
        self._user_id = user_id
        self._username = username
        self._is_admin = is_admin
        self._full_name = full_name
        self._password = password

    @classmethod
    def from_dict(cls, dikt) -> 'UserWithPassword':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The UserWithPassword of this UserWithPassword.  # noqa: E501
        :rtype: UserWithPassword
        """
        return util.deserialize_model(dikt, cls)

    @property
    def user_id(self) -> int:
        """Gets the user_id of this UserWithPassword.


        :return: The user_id of this UserWithPassword.
        :rtype: int
        """
        return self._user_id

    @user_id.setter
    def user_id(self, user_id: int):
        """Sets the user_id of this UserWithPassword.


        :param user_id: The user_id of this UserWithPassword.
        :type user_id: int
        """

        self._user_id = user_id

    @property
    def username(self) -> str:
        """Gets the username of this UserWithPassword.


        :return: The username of this UserWithPassword.
        :rtype: str
        """
        return self._username

    @username.setter
    def username(self, username: str):
        """Sets the username of this UserWithPassword.


        :param username: The username of this UserWithPassword.
        :type username: str
        """

        self._username = username

    @property
    def is_admin(self) -> bool:
        """Gets the is_admin of this UserWithPassword.


        :return: The is_admin of this UserWithPassword.
        :rtype: bool
        """
        return self._is_admin

    @is_admin.setter
    def is_admin(self, is_admin: bool):
        """Sets the is_admin of this UserWithPassword.


        :param is_admin: The is_admin of this UserWithPassword.
        :type is_admin: bool
        """

        self._is_admin = is_admin

    @property
    def full_name(self) -> str:
        """Gets the full_name of this UserWithPassword.


        :return: The full_name of this UserWithPassword.
        :rtype: str
        """
        return self._full_name

    @full_name.setter
    def full_name(self, full_name: str):
        """Sets the full_name of this UserWithPassword.


        :param full_name: The full_name of this UserWithPassword.
        :type full_name: str
        """

        self._full_name = full_name

    @property
    def password(self) -> str:
        """Gets the password of this UserWithPassword.


        :return: The password of this UserWithPassword.
        :rtype: str
        """
        return self._password

    @password.setter
    def password(self, password: str):
        """Sets the password of this UserWithPassword.


        :param password: The password of this UserWithPassword.
        :type password: str
        """

        self._password = password